+++ OpenGL-basico/ray-tracing/cylinder.cpp
@@ -21,65 +21,65 @@ bool cylinder::test_intersection(ray& rayo, vector3& point, vector3& normal)
     vector3 n = axis_; // Eje del cilindro (ya normalizado)
 
     // Proyecciones sobre el eje del cilindro
-    float md = m.dot_product(n);
-    float nd = d.dot_product(n);
+    double md = m.dot_product(n);
+    double nd = d.dot_product(n);
 
     // Coeficiente cuadrático
-    float a = d.dot_product(d) - nd * nd; // d.dot_product(d) = dx^2 + dy^2 + dz^2
+    double a = d.dot_product(d) - nd * nd; // d.dot_product(d) = dx^2 + dy^2 + dz^2
     //Elimino la componente de la dirección del rayo que está alineada con el eje del cilindro.
     //Es útil para determinar si el rayo intersecta el cilindro o no,
     //al considerar solo las componentes del rayo que están perpendiculares al eje del cilindro
 
 
     // Coeficiente constante
-    float c = m.dot_product(m) - md * md - radius_ * radius_; // m.dot_product(m) = mx^2 + my^2 + mz^2
+    double c = m.dot_product(m) - md * md - radius_ * radius_; // m.dot_product(m) = mx^2 + my^2 + mz^2
 
     // Coeficiente lineal
-    float b = 2.0f * (m.dot_product(d) - md * nd);
+    double b = 2.0 * (m.dot_product(d) - md * nd);
 
-    float discriminant = (b * b) - 4 * a * c;
+    double discriminant = (b * b) - 4 * a * c;
 
-    if (discriminant < 0.0f) // Si el discriminante es negativo, no hay intersección
+    if (discriminant < 0.0) // Si el discriminante es negativo, no hay intersección
     {
         return false;
     }
 
-    float sqrt_discriminant = sqrtf(discriminant);
-    float t1 = (-b - sqrt_discriminant) / (2.0f * a);
-    float t2 = (-b + sqrt_discriminant) / (2.0f * a);
+    double sqrt_discriminant = sqrt(discriminant);
+    double t1 = (-b - sqrt_discriminant) / (2.0 * a);
+    double t2 = (-b + sqrt_discriminant) / (2.0 * a);
 
     // Si ambos t1 y t2 son negativos, el cilindro está detrás del origen del rayo
-    if (t2 < 0.0f)
+    if (t2 < 0.0)
     {
         return false;
     }
 
     // Elegimos el menor t positivo
-    float t = t1;
-    if (t < 0.0f)
+    double t = t1;
+    if (t < 0.0)
     {
         t = t2;
     }
 
     vector3 hit_point = rayo.get_origin() + d * t; // Punto de intersección candidato
 
-    float hit_height = (hit_point - get_position()).dot_product(n);
+    double hit_height = (hit_point - get_position()).dot_product(n);
     // Proyección del punto de intersección sobre el eje del cilindro
 
     // Verificación si el punto de intersección está dentro de los límites del cilindro
-    if (hit_height < 0.0f || hit_height > height_)
+    if (hit_height < 0.0 || hit_height > height_)
     {
-        if (hit_height < 0.0f)
+        if (hit_height < 0.0)
         {
             // Verificamos intersección con las tapas del cilindro
-            t = (0.0f - md) / nd;
+            t = (0.0 - md) / nd;
         }
         else
         {
             t = (height_ - md) / nd;
         }
 
-        if (t < 0.0f) // Si t es negativo, la tapa está detrás del origen del rayo
+        if (t < 0.0) // Si t es negativo, la tapa está detrás del origen del rayo
         {
             return false;
         }
@@ -93,7 +93,7 @@ bool cylinder::test_intersection(ray& rayo, vector3& point, vector3& normal)
         }
 
         normal = n;
-        if (hit_height < 0.0f)
+        if (hit_height < 0.0)
         {
             normal = -n;
         }
